+++ tests/testthat/ollama/localhost-11434/api/chat-6c3e3d-POST.R
@@ -1,0 +1,5 @@
+structure(list(method = "POST", url = "http://localhost:11434/api/chat", 
+    status_code = 200L, headers = structure(list(`Content-Type` = "application/json; charset=utf-8", 
+        Date = "Thu, 21 Aug 2025 12:41:21 GMT", `Content-Length` = "330"), redact = character(0), class = "httr2_headers"), 
+    body = charToRaw("{\"model\":\"gemma2\",\"created_at\":\"2025-08-21T12:41:21.294827Z\",\"message\":{\"role\":\"assistant\",\"content\":\"Hello! 👋  How can I help you today?\"},\"done_reason\":\"stop\",\"done\":true,\"total_duration\":9819624917,\"load_duration\":4533008625,\"prompt_eval_count\":26,\"prompt_eval_duration\":4819809375,\"eval_count\":12,\"eval_duration\":462770333}"), 
+    cache = new.env(parent = emptyenv())), class = "httr2_response")