+++ tests/testthat/test_api_ollama.R
@@ -24,7 +24,7 @@ test_that("ollama function constructs a correct request and dry runs it", {
   expect_true("content-type" %in% names(headers))
   
   #Content length is identical in header
-  expect_equal(headers$`content-length`,"147")
+  expect_equal(headers$`content-length`,"148")
 
   # Check that the content-type is JSON
   expect_equal(headers$`content-type`, "application/json")
@@ -32,7 +32,7 @@ test_that("ollama function constructs a correct request and dry runs it", {
   # Now check the body content to ensure the JSON is constructed as expected
   body_json <- request$body |> jsonlite::toJSON() |> as.character()
 
-  expected_json <- "{\"data\":{\"model\":[\"gemma2\"],\"messages\":[{\"role\":[\"user\"],\"content\":[\"Write a poem about porcupines\"]}],\"options\":{\"temperature\":[0],\"num_ctx\":[2048]},\"stream\":[false]},\"type\":[\"json\"],\"content_type\":[\"application/json\"],\"params\":{\"auto_unbox\":[true],\"digits\":[22],\"null\":[\"null\"]}}"
+  expected_json <- "{\"data\":{\"model\":[\"gemma2\"],\"messages\":[{\"role\":[\"user\"],\"content\":[\"Write a poem about porcupines\"]}],\"options\":{\"temperature\":[0],\"num_ctx\":[32768]},\"stream\":[false]},\"type\":[\"json\"],\"content_type\":[\"application/json\"],\"params\":{\"auto_unbox\":[true],\"digits\":[22],\"null\":[\"null\"]}}"
   # Check if the JSON matches the expected JSON
   expect_equal(body_json, expected_json)